% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{c_pal}
\alias{c_pal}
\title{Custom color palette for C colors}
\usage{
c_pal(...)
}
\arguments{
\item{...}{Optional names of specific colors to return}
}
\description{
Returns a named vector of hexadecimal color values.
}
