% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\docType{data}
\name{c_palette}
\alias{c_palette}
\title{Color palette from the 'Carnelian' color theme}
\format{
An object of class \code{character} of length 13.
}
\usage{
c_palette
}
\description{
Color palette from the 'Carnelian' color theme
}
\keyword{datasets}
