% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{c_scale_fill}
\alias{c_scale_fill}
\title{Integration with ggplot2 'scale_fill_manual()' function}
\usage{
c_scale_fill(...)
}
\arguments{
\item{...}{Optional colors to be chosen from the palette}
}
\description{
Integration with ggplot2 'scale_fill_manual()' function
}
