% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{compute_ci}
\alias{compute_ci}
\title{Compute groupwise means and confidence intervals using t-tests}
\usage{
compute_ci(data, ..., value_col, conf_level = 0.95, alternative = "two.sided")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{conf_level}{Confidence level for the t-test (default = 0.95).}

\item{group_col}{A bare (unquoted) grouping variable.}
}
\value{
A tibble with group means and confidence intervals.
}
\description{
Compute groupwise means and confidence intervals using t-tests
}
