% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{compute_group_ci}
\alias{compute_group_ci}
\title{Compute groupwise means and confidence intervals using t-tests}
\usage{
compute_group_ci(data, group_col, value_col, conf_level = 0.95)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{group_col}{A bare (unquoted) grouping variable.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{conf_level}{Confidence level for the t-test (default = 0.95).}
}
\value{
A tibble with group means and confidence intervals.
}
\description{
Compute groupwise means and confidence intervals using t-tests
}
