% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{dlst}
\alias{dlst}
\title{Density function for location-scale t-distribution}
\usage{
dlst(mu, df, x_bar = 0, s_n = 1)
}
\arguments{
\item{mu}{Location parameter.}

\item{df}{Degrees of freedom.}

\item{x}{Vector of quantiles.}

\item{sigma}{Scale parameter (σ > 0).}
}
\value{
Vector of density values.
}
\description{
Density function for location-scale t-distribution
}
