% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{ggplot_box_legend}
\alias{ggplot_box_legend}
\title{Creates a boxplot legend explaining the components of a boxplot}
\usage{
ggplot_box_legend(family = "serif")
}
\arguments{
\item{family}{Font family for the text in the legend.}
}
\value{
A ggplot object with the boxplot legend.
}
\description{
Creates a boxplot legend explaining the components of a boxplot
}
