% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{kmode}
\alias{kmode}
\title{Compute the Statistical Mode}
\usage{
kmode(x)
}
\arguments{
\item{x}{A vector of numeric or character values.}
}
\value{
A single value representing the mode of the input vector.
}
\description{
Returns the mode (most frequent value) of a numeric or character vector.
If there are multiple modes, the first one encountered is returned.
}
\examples{
kmode(c(1, 2, 2, 3, 3, 3, 4))
# Returns: 3

kmode(c("apple", "banana", "apple", "orange"))
# Returns: "apple"

}
