% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils-package.R
\docType{package}
\name{krulRutils-package}
\alias{krulRutils}
\alias{krulRutils-package}
\title{Utility functions for MA2003B}
\description{
Internal helper package used across the MA2003B course materials.
}
\keyword{internal}
