% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Min-Max Scaling of a Numeric Vector}
\usage{
min_max_scale(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector to be scaled.}

\item{na.rm}{Logical, whether to remove NA values before scaling. Default TRUE.}
}
\value{
A numeric vector with the same length as `x`, containing scaled values.
}
\description{
Min-Max Scaling of a Numeric Vector
}
