% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{normal_scale}
\alias{normal_scale}
\title{Normalizing Scaling of a Numeric Vector}
\usage{
normal_scale(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector to be scaled.}

\item{na.rm}{Logical, whether to remove NA values before scaling. Default TRUE.}
}
\value{
A numeric vector with the same length as `x`, of norm one.
}
\description{
Normalizing Scaling of a Numeric Vector
}
