% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{plst}
\alias{plst}
\title{Cumulative distribution function for location-scale t-distribution}
\usage{
plst(q, df, x_bar = 0, s_n = 1)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{df}{Degrees of freedom.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter (σ > 0).}
}
\value{
Vector of probabilities.
}
\description{
Cumulative distribution function for location-scale t-distribution
}
