% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{qlst}
\alias{qlst}
\title{Quantile function for location-scale t-distribution}
\usage{
qlst(p, df, x_bar = 0, s_n = 1)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{df}{Degrees of freedom.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter (σ > 0).}
}
\value{
Vector of quantiles.
}
\description{
Quantile function for location-scale t-distribution
}
