% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{rlst}
\alias{rlst}
\title{Random generation from location-scale t-distribution}
\usage{
rlst(m, df, x_bar = 0, s_n = 1)
}
\arguments{
\item{df}{Degrees of freedom.}

\item{n}{Number of observations.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter (σ > 0).}
}
\value{
Vector of random deviates.
}
\description{
Random generation from location-scale t-distribution
}
