% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{robust_scale}
\alias{robust_scale}
\title{Robust Scaling of a Numeric Vector}
\usage{
robust_scale(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector to be scaled.}

\item{na.rm}{Logical, whether to remove NA values before scaling. Default TRUE.}
}
\value{
A numeric vector with the same length as `x`, containing scaled values.
}
\description{
Robust Scaling of a Numeric Vector
}
