% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{t_test}
\alias{t_test}
\title{Perform a t-test for group means}
\usage{
t_test(data, ..., value_col, mu_0 = 0, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{alpha}{Significance level for the t-test (default = 0.05).}

\item{group_col}{A bare (unquoted) grouping variable.}
}
\value{
A tibble with t-test results
}
\description{
Perform a t-test for group means
}
