% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{theme_krul_ggpairs}
\alias{theme_krul_ggpairs}
\title{Applies the 'krul' theme to a GGally matrix}
\usage{
theme_krul_ggpairs(ggpairs_obj)
}
\arguments{
\item{ggpairs_obj}{A GGally ggpairs object.}
}
\value{
The modified ggpairs object with the 'krul' theme applied.
}
\description{
Applies the 'krul' theme to a GGally matrix
}
