% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{tidy_ad_test}
\alias{tidy_ad_test}
\title{Perform Anderson-Darling normality test for groups}
\usage{
tidy_ad_test(data, ..., value_col)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{group_col}{A bare (unquoted) grouping variable.}
}
\value{
A tibble with Anderson-Darling test results
}
\description{
Perform Anderson-Darling normality test for groups
}
