% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{tidy_jb_test}
\alias{tidy_jb_test}
\title{Perform Jarque-Bera normality test for groups}
\usage{
tidy_jb_test(data, ..., value_col)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{group_col}{A bare (unquoted) grouping variable.}
}
\value{
A tibble with Jarque-Bera test results
}
\description{
Perform Jarque-Bera normality test for groups
}
