% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{tidy_ks_test}
\alias{tidy_ks_test}
\title{Perform Kolgomorov-Smirnov normality test for groups}
\usage{
tidy_ks_test(data, ..., value_col)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{group_col}{A bare (unquoted) grouping variable.}
}
\value{
A tibble with Kolgomorov-Smirnov test results
}
\description{
Perform Kolgomorov-Smirnov normality test for groups
}
