% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krulRutils.R
\name{tidy_shapiro_test}
\alias{tidy_shapiro_test}
\title{Perform Shapiro-Wilk normality test for groups}
\usage{
tidy_shapiro_test(data, ..., value_col)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{value_col}{A bare (unquoted) numeric variable.}

\item{group_col}{A bare (unquoted) grouping variable.}
}
\value{
A tibble with Shapiro-Wilk test results
}
\description{
Perform Shapiro-Wilk normality test for groups
}
